+++ services/horizon/internal/db2/history/ingestion_test.go
@@ -1,0 +1,122 @@
+package history
+
+import (
+	"testing"
+	"time"
+
+	sq "github.com/Masterminds/squirrel"
+	"github.com/guregu/null"
+	"github.com/stellar/go/services/horizon/internal/test"
+	"github.com/stellar/go/services/horizon/internal/toid"
+)
+
+func TestRemoveExpIngestHistory(t *testing.T) {
+	tt := test.Start(t)
+	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
+	q := &Q{tt.HorizonSession()}
+
+	summary, err := q.RemoveExpIngestHistory(69859)
+	tt.Assert.Equal(ExpIngestRemovalSummary{}, summary)
+	tt.Assert.NoError(err)
+
+	txInsertBuilder := q.NewTransactionBatchInsertBuilder(0)
+
+	ledger := Ledger{
+		Sequence:                   69859,
+		PreviousLedgerHash:         null.NewString("4b0b8bace3b2438b2404776ce57643966855487ba6384724a3c664c7aa4cd9e4", true),
+		ImporterVersion:            321,
+		TransactionCount:           12,
+		SuccessfulTransactionCount: new(int32),
+		FailedTransactionCount:     new(int32),
+		OperationCount:             23,
+		TotalCoins:                 23451,
+		FeePool:                    213,
+		BaseReserve:                687,
+		MaxTxSetSize:               345,
+		ProtocolVersion:            12,
+		BaseFee:                    100,
+		ClosedAt:                   time.Now().UTC().Truncate(time.Second),
+		LedgerHeaderXDR:            null.NewString("temp", true),
+	}
+	*ledger.SuccessfulTransactionCount = 12
+	*ledger.FailedTransactionCount = 3
+	hashes := []string{
+		"4db1e4f145e9ee75162040d26284795e0697e2e84084624e7c6c723ebbf80118",
+		"5db1e4f145e9ee75162040d26284795e0697e2e84084624e7c6c723ebbf80118",
+		"6db1e4f145e9ee75162040d26284795e0697e2e84084624e7c6c723ebbf80118",
+		"7db1e4f145e9ee75162040d26284795e0697e2e84084624e7c6c723ebbf80118",
+		"8db1e4f145e9ee75162040d26284795e0697e2e84084624e7c6c723ebbf80118",
+	}
+
+	for i, hash := range hashes {
+		ledger.TotalOrderID.ID = toid.New(ledger.Sequence, 0, 0).ToInt64()
+		ledger.LedgerHash = hash
+		if i > 0 {
+			ledger.PreviousLedgerHash = null.NewString(hashes[i-1], true)
+		}
+
+		insertSQL := sq.Insert("exp_history_ledgers").SetMap(ledgerToMap(ledger))
+		_, err = q.Exec(insertSQL)
+		tt.Assert.NoError(err)
+
+		err = txInsertBuilder.Add(
+			buildLedgerTransaction(
+				tt,
+				testTransaction{
+					index:         1,
+					envelopeXDR:   "AAAAACiSTRmpH6bHC6Ekna5e82oiGY5vKDEEUgkq9CB//t+rAAAAyAEXUhsAADDRAAAAAAAAAAAAAAABAAAAAAAAAAsBF1IbAABX4QAAAAAAAAAA",
+					resultXDR:     "AAAAAAAAASwAAAAAAAAAAwAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAFAAAAAAAAAAA=",
+					feeChangesXDR: "AAAAAA==",
+					metaXDR:       "AAAAAQAAAAAAAAAA",
+					hash:          "19aaa18db88605aedec04659fb45e06f240b022eb2d429e05133e4d53cd945ba",
+				},
+			),
+			uint32(ledger.Sequence),
+		)
+		tt.Assert.NoError(err)
+		tt.Assert.NoError(txInsertBuilder.Exec())
+
+		ledger.Sequence++
+	}
+
+	var ledgers []Ledger
+	err = q.Select(&ledgers, selectLedgerFields.From("exp_history_ledgers hl"))
+	tt.Assert.NoError(err)
+	tt.Assert.Len(ledgers, 5)
+
+	var transactions []Transaction
+	err = q.Select(&transactions, selectExpTransaction)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(transactions, 5)
+
+	summary, err = q.RemoveExpIngestHistory(69863)
+	tt.Assert.Equal(ExpIngestRemovalSummary{}, summary)
+	tt.Assert.NoError(err)
+
+	err = q.Select(&ledgers, selectLedgerFields.From("exp_history_ledgers hl"))
+	tt.Assert.NoError(err)
+	tt.Assert.Len(ledgers, 5)
+
+	err = q.Select(&transactions, selectExpTransaction)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(transactions, 5)
+
+	cutoffSequence := 69861
+	summary, err = q.RemoveExpIngestHistory(uint32(cutoffSequence))
+	tt.Assert.Equal(ExpIngestRemovalSummary{LedgersRemoved: 2, TransactionsRemoved: 2}, summary)
+	tt.Assert.NoError(err)
+
+	err = q.Select(&ledgers, selectLedgerFields.From("exp_history_ledgers hl"))
+	tt.Assert.NoError(err)
+	tt.Assert.Len(ledgers, 3)
+
+	err = q.Select(&transactions, selectExpTransaction)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(transactions, 3)
+
+	for i := range ledgers {
+		tt.Assert.LessOrEqual(ledgers[i].Sequence, int32(cutoffSequence))
+		tt.Assert.LessOrEqual(transactions[i].LedgerSequence, int32(cutoffSequence))
+	}
+}
